requirements.txt
djangorestframework==3.14.0
celery==5.2.7
ipython==8.11.0
Django==3.2
django-celery-beat==2.4.0
requests==2.28.1
redis==4.5.5
# Python==3.10.6
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary
python-env==1.0.0
